refactor(codex): deduplicate command display text
