+++ extensions/codex/src/command-diagnostics-support.ts
@@ -8,6 +8,7 @@ import {
   type CodexDiagnosticsTarget,
   type PendingCodexDiagnosticsConfirmation,
 } from "./command-diagnostics-state.js";
+import { escapeCodexChatText, formatCodexTextForDisplay } from "./command-display-text.js";
 import { splitArgs } from "./command-handler-args.js";
 
 type ParsedDiagnosticsArgs =
@@ -331,33 +332,6 @@ function formatCodexDiagnosticsTargetLine(target: CodexDiagnosticsTarget): strin
   return `- ${parts.join(", ")}`;
 }
 
-export function escapeCodexChatText(value: string): string {
-  return value
-    .replaceAll("&", "&amp;")
-    .replaceAll("<", "&lt;")
-    .replaceAll(">", "&gt;")
-    .replaceAll("@", "\uff20")
-    .replaceAll("`", "\uff40")
-    .replaceAll("[", "\uff3b")
-    .replaceAll("]", "\uff3d")
-    .replaceAll("(", "\uff08")
-    .replaceAll(")", "\uff09")
-    .replaceAll("*", "\u2217")
-    .replaceAll("_", "\uff3f")
-    .replaceAll("~", "\uff5e")
-    .replaceAll("|", "\uff5c");
-}
-
-export function formatCodexTextForDisplay(value: string): string {
-  let safe = "";
-  for (const character of value) {
-    const codePoint = character.codePointAt(0);
-    safe += codePoint != null && isUnsafeDisplayCodePoint(codePoint) ? "?" : character;
-  }
-  safe = safe.trim();
-  return safe || "<unknown>";
-}
-
 export function readCodexDiagnosticsTargetsCooldownMessage(
   targets: readonly CodexDiagnosticsTarget[],
   ctx: PluginCommandContext,
@@ -524,22 +498,6 @@ function formatCodexResumeCommandForDisplay(threadId: string): string {
     return "run codex resume and paste the thread id shown above";
   }
   return `\`codex resume ${safeThreadId}\``;
-}
-
-function isUnsafeDisplayCodePoint(codePoint: number): boolean {
-  return (
-    codePoint <= 0x001f ||
-    (codePoint >= 0x007f && codePoint <= 0x009f) ||
-    codePoint === 0x00ad ||
-    codePoint === 0x061c ||
-    codePoint === 0x180e ||
-    (codePoint >= 0x200b && codePoint <= 0x200f) ||
-    (codePoint >= 0x202a && codePoint <= 0x202e) ||
-    (codePoint >= 0x2060 && codePoint <= 0x206f) ||
-    codePoint === 0xfeff ||
-    (codePoint >= 0xfff9 && codePoint <= 0xfffb) ||
-    (codePoint >= 0xe0000 && codePoint <= 0xe007f)
-  );
 }
 
 function normalizeCodexDiagnosticsScopeField(value: string | undefined): string | undefined {
